requirements.txt
aiogram
sqlalchemy
python-dotenv
openai
psycopg[binary]
fastapi
uvicorn